Module polyfuse_examples::prelude

Re-exports

pub use crate as examples;

Macros

anyhow
Construct an ad-hoc error from a string.

ensure
Return early with an error if a condition is not satisfied.

Structs

Context
Contextural information about an incoming request.

OsStr
Borrowed reference to an OS string (see OsString).

OsString
A type that can represent owned, mutable platform-native strings, but is cheaply inter-convertible with Rust strings.

Path
A slice of a path (akin to str).

PathBuf
An owned, mutable path (akin to String).

Enums

Operation
The kind of FUSE requests received from the kernel.

Traits

AsyncWrite
Write bytes asynchronously.

Filesystem
The filesystem running on the user space.

Future
A future represents an asynchronous computation.

FutureExt
An extension trait for Futures that provides a variety of convenient adapters.

OsStrExt
Platform-specific extensions to OsStr.

Attribute Macros

async_trait